--- Assets/AudioTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b934f33513963ea4495767ef3166d115
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
