+++ Assets/NavMeshComponents/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 63b588f3892bb4b5eb73ad3d2791e05c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
