+++ Assets/NavMeshComponents/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ce67aa87f613246dda63a54a59c6399e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
