--- Assets/PlayerInteraction.cs
@@ -1,5 +1,0 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
