+++ Assets/Resources/Map/Tilemap_Palette/TestItemPalette.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0897a9a0e843ddc4c8e8a77b2586060d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
