+++ Assets/Scenes/UI_Game.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b1eb190c309a1db4eabd913b5db55647
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
