--- Assets/Script.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ae6bd335eaf3f3a43801fb53cecac526
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
