--- Assets/Script/TimeInGame.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1d15a33400c7d9346ac7af153fe5e372
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
