--- Assets/Scripts/InnerMapTransfer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 07dca58a34c31f64eaaff4a06e3cbe75
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
