+++ Assets/Scripts/Item/ItemMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 73c1371bf72426342bad038ae8f33237
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
