+++ Assets/Scripts/Map/MapCoordinate.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9dac5f7e2afbd8f43a6f8595bd10e7af
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
