+++ Assets/Scripts/Player/PlayerInteraction.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eb13c262784ed8a42b04c76fc178eb5f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
