+++ Assets/Scripts/Quest_General/QuestManager.cs
@@ -7,32 +7,28 @@ using System.IO;
 using System.Linq;
 
 public static class QuestManager{
-    static private QuestDatabase database = null;
     static DS.UI.Window questWindow;        //DetailのWindowAcrivatorのInactivateWindow(キャンセルしたときに消えるWindow)に設定する
     static DS.UI.Tab questTab;              //HeaderとDetailを動的生成した後にLinkTabHeaderを呼ぶため
    
     public static Dictionary<string, Quest_Behaviour> OrderedQuest {get; private set; }     //クエスト名をキーとし,受注済みクエストを値とする辞書
     public static Dictionary<string,Quest_Behaviour> ClearedQuest { get; private set; }     //クエスト名をキーとし,クリア済みクエストを値とする辞書
-    public static void Init(Transform headerParent,DS.UI.Tab tab) {//QuestManagerBehaviorのStartで呼び出される
-        questTab = tab;
+    public static void Init() {//QuestManagerBehaviorのStartで呼び出される
         OrderedQuest = new Dictionary<string, Quest_Behaviour>();
         ClearedQuest = new Dictionary<string, Quest_Behaviour>();
-        database = new QuestDatabase();
+        QuestDatabase.Init();
     }
-    static public void QuestAccept(string questName){
-        if (!database.AcceptableQuest.ContainsKey(questName)) {
+    static public void QuestAccept(string questName) {
+        if (!QuestDatabase.AcceptableQuest.ContainsKey(questName)) {
             Debug.LogError(questName + "が受注可能クエスト内に見つかりません");
             return;
         }
-        Quest_Behaviour quest = database.AcceptableQuest[questName];
-        database.AcceptableQuest.Remove(questName);
+        Quest_Behaviour quest = QuestDatabase.AcceptableQuest[questName];
+        QuestDatabase.AcceptableQuest.Remove(questName);
         //quest.WhenQuestAccepted?.Invoke();
         var obj=GameObject.Instantiate(quest);
         obj.name = questName;//名前から(Clone)を除く
         OrderedQuest.Add(questName, obj);
-        QuestHeaderGenerator.Generate(database.questHeaderPrefab, quest.info.displayName);
-        QuestDetailGenerator.Generate(database.questDetailPrefab, quest.info);
-        questTab.LinkTabHeader();
+        QuestMenu.AddNewQuestUI(quest.info);
     }    
     static public void QuestClear(string questName){
         if (OrderedQuest.ContainsKey(questName)) {
@@ -43,7 +39,7 @@ public static class QuestManager{
             ClearedQuest.Add(questName, quest);
 
             GameObject.Destroy(quest);
-            QuestHeaderGenerator.EnCleared(quest.info.displayName);
+            QuestMenu.EnCleared(quest.info.displayName);
         }
         else {
             Debug.LogError("受注済みクエストに" + questName + "がない状態でQuestClearが呼ばれました");
@@ -64,81 +60,14 @@ public static class QuestManager{
     }
 }
 
-public static class QuestHeaderGenerator {
-    static Transform parentTransform;
-    static Dictionary<string, GameObject> nowValidHeader = null;
-    const int headerPosX = 280;
-    const int headerOffset = 20;
-    public static void Init(Transform parent) {
-        parentTransform = parent;
-        nowValidHeader = new Dictionary<string, GameObject>();
-    }
-    public static void Generate(GameObject prefab, string displayName) {
-        GameObject obj = GameObject.Instantiate(prefab, parentTransform);
-        var text = obj.GetComponentInChildren<Text>();
-        text.text = displayName;
-        var rect = obj.GetComponent<RectTransform>();
-        var array = nowValidHeader.ToArray();
-        float topPosY = -120;
-        if (array.Length > 0) { //既にあるheaderの中で一番上の座標を見て、その上になるように
-            float minY = float.PositiveInfinity;
-            foreach(float posY in nowValidHeader.Values.Select(x => x.GetComponent<RectTransform>().sizeDelta.y)) {
-                minY = Mathf.Min(minY, posY);
-            }
-            topPosY = minY;
-        }
-        rect.anchoredPosition = new Vector2(headerPosX, topPosY + (rect.sizeDelta.y + headerOffset));
-        nowValidHeader.Add(displayName, obj);
-    }
-    public static void EnCleared(string displayName) {
-        nowValidHeader[displayName].GetComponent<Image>().color = Color.black;
-    }
-}
-public static class QuestDetailGenerator{
-    static Transform parentTransform;
-    public static void Init(Transform parent) {
-        parentTransform = parent;
-    }
-    public static void Generate(GameObject prefab,QuestInfo info) {
-        GameObject obj = GameObject.Instantiate(prefab, parentTransform);
-        var detailBehaviour = obj.GetComponent<QuestDetail_Behaviour>();
-        if (detailBehaviour == null) {
-            Debug.LogError("QuestDetail_Behaviour.GenerateDetailで、QuestDetail_Behaivourがアタッチされていないprefabを生成しようとしました");
-            return;
-        }
-        var textList = obj.GetComponentsInChildren<UnityEngine.UI.Text>();
-        foreach (var text in textList) {
-            switch (text.gameObject.name) {
-                case "name":
-                    text.text = info.displayName;
-                    break;
-                case "target":
-                    text.text = info.target;
-                    break;
-                case "place":
-                    text.text = info.place;
-                    break;
-                case "reward":
-                    text.text = info.reward;
-                    break;
-                case "detail":
-                    text.text = info.detail;
-                    break;
-                default: break;
-            }
-        }
-        detailBehaviour.RemainingTime = info.deadLine - TimeInGame.Current;
-    }
-}
-
-public class QuestDatabase {
-    public GameObject questHeaderPrefab;
-    public GameObject questDetailPrefab;
-    public Dictionary<string, Quest_Behaviour> AcceptableQuest { get; private set; } //クエスト名をキーとし,受注可能クエストを値とする辞書
-    public Dictionary<string, Sprite> QuestHeaderSprite { get; private set; } //クエスト名をキーとし、メニューで表示されるヘッダーの画像を値とする辞書
-    public Dictionary<string, Sprite> QuestDetailSprite { get; private set; } //クエスト名をキーとし、メニューで表示される詳細の画像を値とする辞書
+public static class QuestDatabase {
+    public static GameObject questHeaderPrefab;
+    public static GameObject questDetailPrefab;
+    public static Dictionary<string, Quest_Behaviour> AcceptableQuest { get; private set; } //クエスト名をキーとし,受注可能クエストを値とする辞書
+    public static Dictionary<string, Sprite> QuestHeaderSprite { get; private set; } //クエスト名をキーとし、メニューで表示されるヘッダーの画像を値とする辞書
+    public static Dictionary<string, Sprite> QuestDetailSprite { get; private set; } //クエスト名をキーとし、メニューで表示される詳細の画像を値とする辞書
     
-    public QuestDatabase() {
+    public static void Init() {
         string path = @"Quest/UI/Prefabs/";
         questHeaderPrefab = Resources.Load<GameObject>(path + "questHeader");
         questDetailPrefab = Resources.Load<GameObject>(path + "questDetail");
@@ -146,7 +75,7 @@ public class QuestDatabase {
         InitQuestSprite();
     }
 
-    void InitAcceptableQuest() {
+    static void InitAcceptableQuest() {
         AcceptableQuest = new Dictionary<string, Quest_Behaviour>();
         string path = @"Quest/IndividualPrefabs/";
         var prefabs = Resources.LoadAll<GameObject>(path);
@@ -154,7 +83,7 @@ public class QuestDatabase {
             AcceptableQuest[p.name] = p.GetComponent<Quest_Behaviour>();
         }
     }
-    void InitQuestSprite() {
+    static void InitQuestSprite() {
         QuestHeaderSprite = new Dictionary<string, Sprite>();
         string path = @"Quest/UI/HeaderImages";
         var headerSprites = Resources.LoadAll<Sprite>(path);
