--- Assets/Scripts/System.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 04f933fe6320e614c96c998af4affc2c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
