--- Assets/Scripts/System_Behaviour.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0ff8a448ea63ffb4caf6d05689dc5aec
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
