+++ Assets/Scripts/UI/MenuInUIScene_Behaviour.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a95bc0fda432c7045bfc7f15bbf016e5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
