+++ Assets/Scripts/UI/QuestMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 370ce66f3ea0bc44a9f91150205f53fb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
